Move citool.go to citool/main.go
This change should hpefully fix the issue where
`go get https://github.com/ashishb/ci-analysis-tool` creates
`ci-analysis-tool` binary since that's the last dir name in the path.
