+++ citool/src/citool/analyze.go
