+++ citool/src/citool/filter.go
