+++ citool/src/citool/jobstatus.go
