+++ citool/src/citool/utils.go
