+++ test.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-GO111MODULE=on go run citool.go --mode analyze  --print-duration-graph=false --print-success-graph=false --input-files test/circleci_data/*.json > test/analyze_actual_output.txt
+GO111MODULE=on go run citool/main.go --mode analyze  --print-duration-graph=false --print-success-graph=false --input-files test/circleci_data/*.json > test/analyze_actual_output.txt
 diff test/analyze_actual_output.txt test/analyze_expected_output.txt
 rm test/analyze_actual_output.txt
